trading-bot-backend/requirements.txt
python-binance>=1.0.19
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
Flask>=3.0.0
Flask-CORS>=4.0.0
pandas>=2.0.0
gunicorn>=21.2.0
requests>=2.31.0 